refactor: consolidate WranglerDevServer dependencies and improve test structure
- Removed unnecessary dependencies on `@livestore/utils-dev` in `package.json` and `tsconfig.json` for the CLI package.
- Updated test files to utilize the new `WranglerDevServerService` import path, enhancing clarity and maintainability.
- Streamlined the `pnpm-lock.yaml` by removing outdated references and adding new dependencies for improved package management.
- Introduced a new `wrangler` module to encapsulate `WranglerDevServer` functionalities, promoting better organization and modularity.
