--- packages/@livestore/utils-dev/src/node/WranglerDevServer/WranglerDevServer.test.ts
@@ -1,270 +1,0 @@
-import { Effect, Exit, FetchHttpClient, Fiber, Layer, Scope } from '@livestore/utils/effect'
-import { getFreePort, PlatformNode } from '@livestore/utils/node'
-import { Vitest } from '@livestore/utils-dev/node-vitest'
-import { expect } from 'vitest'
-import {
-  type StartWranglerDevServerArgs,
-  WranglerDevServerError,
-  WranglerDevServerService,
-} from './WranglerDevServer.ts'
-
-const testTimeout = 60_000
-
-const withTestCtx = Vitest.makeWithTestCtx({
-  timeout: testTimeout,
-  makeLayer: () => PlatformNode.NodeContext.layer,
-})
-
-const WranglerDevServerTest = (args: Partial<StartWranglerDevServerArgs> = {}) =>
-  WranglerDevServerService.Default({
-    cwd: `${import.meta.dirname}/fixtures`,
-    ...args,
-  }).pipe(Layer.provide(FetchHttpClient.layer))
-
-Vitest.describe('WranglerDevServer', { timeout: testTimeout }, () => {
-  Vitest.describe('Basic Operations', () => {
-    const withBasicTest = (args: Partial<StartWranglerDevServerArgs> = {}) =>
-      Vitest.makeWithTestCtx({
-        timeout: testTimeout,
-        makeLayer: () => WranglerDevServerTest(args).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-      })
-
-    Vitest.scopedLive('should start wrangler dev server and return port', (test) =>
-      Effect.gen(function* () {
-        const server = yield* WranglerDevServerService
-
-        expect(server.port).toBeGreaterThan(0)
-        expect(server.url).toMatch(/http:\/\/localhost:\d+/)
-        expect(typeof server.processId).toBe('number')
-        expect(server.processId).toBeGreaterThan(0)
-      }).pipe(withBasicTest()(test)),
-    )
-
-    Vitest.scopedLive('should use specified port when provided', (test) =>
-      Effect.andThen(getFreePort, (port) =>
-        Effect.gen(function* () {
-          const server = yield* WranglerDevServerService
-
-          expect(server.port).toBe(port)
-          expect(server.url).toBe(`http://localhost:${port}`)
-        }).pipe(withBasicTest({ preferredPort: port })(test)),
-      ),
-    )
-  })
-
-  Vitest.describe('Resource Management', () => {
-    Vitest.scopedLive('should cleanup processes on scope close', (test) =>
-      Effect.gen(function* () {
-        let processId: number | undefined
-
-        // Create a separate scope for the server
-        const serverScope = yield* Scope.make()
-
-        const server = yield* Effect.provide(
-          WranglerDevServerService,
-          WranglerDevServerTest().pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-        ).pipe(Scope.extend(serverScope))
-
-        processId = server.processId
-        expect(processId).toBeGreaterThan(0)
-        expect(server.port).toBeGreaterThan(0)
-        expect(server.url).toMatch(/http:\/\/localhost:\d+/)
-
-        // Close scope to trigger cleanup
-        yield* Scope.close(serverScope, Exit.succeed(void 0))
-
-        // Wait for cleanup to complete
-        yield* Effect.sleep('2 seconds')
-
-        // Verify process is terminated
-        const isRunning2 = yield* Effect.promise(() => {
-          try {
-            process.kill(processId!, 0)
-            return Promise.resolve(true)
-          } catch {
-            return Promise.resolve(false)
-          }
-        })
-        expect(isRunning2).toBe(false)
-      }).pipe(withTestCtx(test)),
-    )
-
-    Vitest.scopedLive('should handle interruption with fast cleanup', (test) =>
-      Effect.gen(function* () {
-        let processId: number | undefined
-
-        const fiber = yield* Effect.fork(
-          Effect.provide(
-            Effect.gen(function* () {
-              const server = yield* WranglerDevServerService
-              processId = server.processId
-              yield* Effect.sleep('30 seconds') // Keep running
-              return server
-            }),
-            WranglerDevServerTest().pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-          ),
-        )
-
-        // Wait for server to start
-        yield* Effect.sleep('3 seconds')
-
-        expect(processId).toBeGreaterThan(0)
-
-        // Interrupt and measure cleanup time
-        const start = Date.now()
-        yield* Fiber.interrupt(fiber)
-        const elapsed = Date.now() - start
-
-        // Should use fast cleanup (500ms timeout) + some overhead
-        expect(elapsed).toBeLessThan(1500) // Allow some overhead
-
-        // Wait for cleanup to complete
-        yield* Effect.sleep('1 second')
-
-        // Verify process is terminated
-        const isRunningAfter = yield* Effect.promise(() => {
-          try {
-            process.kill(processId!, 0)
-            return Promise.resolve(true)
-          } catch {
-            return Promise.resolve(false)
-          }
-        })
-        expect(isRunningAfter).toBe(false)
-      }).pipe(withTestCtx(test)),
-    )
-  })
-
-  Vitest.describe('Error Handling', () => {
-    const withErrorTest = (args: Partial<StartWranglerDevServerArgs> = {}) =>
-      Vitest.makeWithTestCtx({
-        timeout: testTimeout,
-        makeLayer: () => WranglerDevServerTest(args).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-      })
-
-    Vitest.scopedLive('should handle missing wrangler.toml but should timeout', (test) =>
-      Effect.gen(function* () {
-        const error = yield* WranglerDevServerService.pipe(
-          Effect.provide(
-            WranglerDevServerTest({
-              cwd: '/tmp',
-              wranglerConfigPath: '/dev/null',
-              connectTimeout: '500 millis',
-            }).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-          ),
-          Effect.flip,
-        )
-
-        expect(error).toBeInstanceOf(WranglerDevServerError)
-      }).pipe(Vitest.withTestCtx(test)),
-    )
-
-    Vitest.scopedLive('should handle invalid working directory', (test) =>
-      Effect.gen(function* () {
-        const result = yield* WranglerDevServerService.pipe(
-          Effect.provide(
-            WranglerDevServerTest({
-              cwd: '/completely/nonexistent/directory',
-            }).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-          ),
-          Effect.either,
-        )
-
-        expect(result._tag).toBe('Left')
-        if (result._tag === 'Left') {
-          expect(result.left).toBeInstanceOf(WranglerDevServerError)
-        }
-      }).pipe(Vitest.withTestCtx(test)),
-    )
-
-    Vitest.scopedLive('should timeout if server fails to start', (test) =>
-      Effect.gen(function* () {
-        // Create a command that will never output "Ready on"
-        const result = yield* WranglerDevServerService.pipe(
-          // Override the timeout for this test to be shorter
-          Effect.timeout('5 seconds'),
-          Effect.either,
-        )
-
-        // This might succeed or fail depending on actual wrangler behavior
-        // The main point is testing timeout functionality
-        expect(['Left', 'Right']).toContain(result._tag)
-      }).pipe(withErrorTest()(test)),
-    )
-  })
-
-  Vitest.describe('Process Tree Cleanup', () => {
-    const withCleanupTest = (args: Partial<StartWranglerDevServerArgs> = {}) =>
-      Vitest.makeWithTestCtx({
-        timeout: testTimeout,
-        makeLayer: () => WranglerDevServerTest(args).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-      })
-
-    Vitest.scopedLive('should clean up child workerd processes', (test) =>
-      Effect.gen(function* () {
-        let processId: number | undefined
-
-        const server = yield* WranglerDevServerService
-        processId = server.processId
-
-        // Wait for wrangler to spawn workerd children
-        yield* Effect.sleep('3 seconds')
-
-        // Find any child processes (workerd)
-        const children = yield* Effect.promise(async () => {
-          const { exec } = require('node:child_process')
-          const { promisify } = require('node:util')
-          const execAsync = promisify(exec)
-
-          try {
-            if (!processId) throw new Error('processId is undefined')
-            const { stdout } = await execAsync(`ps -o pid,ppid -ax | grep -E "^\\s*[0-9]+\\s+${processId}\\s*$"`)
-            return stdout
-              .trim()
-              .split('\n')
-              .map((line: string) => {
-                const match = line.trim().match(/^\s*(\d+)\s+\d+\s*$/)
-                return match?.[1] ? Number.parseInt(match[1], 10) : null
-              })
-              .filter((pid: number | null): pid is number => pid !== null)
-          } catch {
-            return []
-          }
-        })
-
-        console.log(`Found ${children.length} child processes:`, children)
-
-        // The scope will close here and should clean up all processes
-      }).pipe(withCleanupTest()(test)),
-    )
-  })
-
-  Vitest.describe('Service Pattern', () => {
-    const withServiceTest = (args: Partial<StartWranglerDevServerArgs> = {}) =>
-      Vitest.makeWithTestCtx({
-        timeout: testTimeout,
-        makeLayer: () => WranglerDevServerTest(args).pipe(Layer.provide(PlatformNode.NodeContext.layer)),
-      })
-
-    Vitest.scopedLive('should work with service pattern', (test) =>
-      Effect.gen(function* () {
-        const server = yield* WranglerDevServerService
-
-        expect(server.port).toBeGreaterThan(0)
-        expect(server.url).toMatch(/http:\/\/localhost:\d+/)
-        expect(server.processId).toBeGreaterThan(0)
-      }).pipe(withServiceTest()(test)),
-    )
-
-    Vitest.scopedLive('should work with custom port via service', (test) =>
-      Effect.andThen(getFreePort, (port) =>
-        Effect.gen(function* () {
-          const server = yield* WranglerDevServerService
-
-          expect(server.port).toBe(port)
-          expect(server.url).toBe(`http://localhost:${port}`)
-        }).pipe(withServiceTest({ preferredPort: port })(test)),
-      ),
-    )
-  })
-})
