+++ tests/integration/src/tests/node-sync/vitest.config.ts
@@ -1,7 +1,5 @@
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
-  test: {
-    setupFiles: ['./fixtures/vitest-sync-setup.ts'],
-  },
+  test: {},
 })
